fix(deploy): separated apis
